// src/types.ts
export interface PinpointDetailsResponse {
    LocationID: string;
    BusinessName: string | null;
    HouseNumber: string | null;
    PremiseName: string | null;
    PremiseLaneName: string | null;
    Moo: string | null;
    StreetLeadingType: string | null;
    StreetName: string | null;
    StreetTrailingType: string | null;
    SubStreetLeadingType: string | null;
    SubStreetName: string | null;
    SubStreetTrailingType: string | null;
    StreetFullName: string | null;
    SubDistrictPrefix: string | null;
    SubDistrict: string;
    DistrictPrefix: string | null;
    District: string;
    ProvincePrefix: string | null;
    Province: string;
    LanguageCode: string;
    LAT_LON: string;
    PostalCode: string;
    FormattedAddress: string;
}

/**
 * ข้อมูลที่ได้จากการค้นหาสถานที่ด้วย Pinpoint Autocomplete ประกอบด้วยข้อมูลพื้นฐานของสถานที่
 * @property LocationID รหัสสถานที่
 * @property FormattedAddress ที่อยู่ที่จัดรูปแบบแล้ว
 */
export interface PinpointAutocompleteResponse {
    LocationID: string;
    FormattedAddress: string;
}
// ...
export interface PinpointBatchResponse {
    success: boolean;
    fields: keyof PinpointDetailsResponse[];
    data: {
        input: string;
        result: {
            type: "address" | "point_of_interest" | "business";
            grade: "A" | "B+" | "B" | "C" | "D" | "E" | "F" | "-";
            match: PinpointDetailsResponse[];
            score: number;
        };
    }[];
}

/**
 * ข้อมูลการตั้งค่าของ Pinpoint Service
 * @property token โทเคนสำหรับใช้งาน API ของ Pinpoint
 * @property referer ที่อยู่ URL ของเว็บไซต์ที่ใช้งาน API ของ Pinpoint
 * @property baseUrl ที่อยู่ URL ของ API ของ Pinpoint (ค่าดีฟอลต์คือ https://pin-point.co/g/search/)
 */
export interface PinpointServiceSettings {
    token: string;
    referer: string;
    baseUrl?: string; // Optional, with a default value
}
// ...
export interface IPinpointService {
    Autocomplete(
        location: string,
        maxResult?: number,
    ): Promise<PinpointAutocompleteResponse[]>;
    Details(location: string): Promise<PinpointDetailsResponse | null>;
    BatchDetails(location: string[]): Promise<PinpointBatchResponse | null>;
}